GetTicksAndSaveDollarBars/requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
multitasking==0.0.9
numpy==1.20.3
pandas==1.2.4
pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.26.4
yfinance==0.1.59
google-cloud-storage>=1.35.0
google-cloud-bigquery==2.16.1
pyarrow>=3.0.0
